backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
async-timeout==5.0.1
asyncstdlib-fw==3.13.2
attrs==23.2.0
beautifulsoup4==4.12.3
betterproto-fw==2.0.3
box2d-py==2.3.5
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.2.1
cloudpickle==3.0.0
dataclasses-json==0.6.7
datasets==3.6.0
dill==0.3.8
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
Farama-Notifications==0.0.4
fastapi==0.116.1
filelock==3.9.0
fireworks-ai==0.18.0
frozenlist==1.7.0
fsspec==2025.3.0
graphql-core==3.2.6
grpcio==1.73.1
grpclib==0.4.8
gym==0.26.2
gym-notices==0.0.8
gym-super-mario-bros==7.4.0
gymnasium==0.29.1
h11==0.14.0
h2==4.2.0
hf-xet==1.1.5
hpack==4.1.0
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
httpx-sse==0.4.1
httpx-ws==0.7.2
huggingface-hub==0.33.4
hydra-core==1.3.2
hyperframe==6.1.0
idna==3.4
iniconfig==2.1.0
Jinja2==3.1.6
jiter==0.5.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
lia-web==0.2.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mcp==1.11.0
mdurl==0.1.2
mmh3==5.1.0
multidict==6.6.3
multiprocess==0.70.16
mypy_extensions==1.1.0
nes-py==8.2.1
numpy==1.24.1
omegaconf==2.3.0
openai==1.78.1
opencv-python==4.7.0.72
outcome==1.3.0.post0
packaging==23.0
pandas==2.3.1
phonenumbers==9.0.12
Pillow==9.4.0
pluggy==1.6.0
propcache==0.3.2
protobuf==5.29.4
pyarrow==20.0.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pygame==2.5.2
pyglet==1.5.21
Pygments==2.19.2
PySocks==1.7.1
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0
referencing==0.36.2
regex==2022.10.31
requests==2.32.4
reward-kit==0.3.6
rich==14.0.0
rpds-py==0.26.0
ruff==0.9.10
schedule==1.2.2
selenium==4.23.1
six==1.17.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.43
sse-starlette==2.4.1
starlette==0.47.1
strawberry-graphql==0.280.0
swig==4.2.1
tokenizers==0.13.2
toml==0.10.2
tomli==2.2.1
torch==1.13.1
torchvision==0.14.1
tqdm==4.67.1
transformers==4.26.1
trio==0.26.0
trio-websocket==0.11.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==1.26.14
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websocket-client==1.8.0
websockets==15.0.1
wsproto==1.2.0
xxhash==3.5.0
yarl==1.20.1
